create call-tokenuri factory and use it where ever applicable
